respark_backend/requirements.txt
Flask==2.2.2
Flask_Cors==5.0.0
lida==0.0.14
matplotlib==3.8.2
numpy==1.24.3
openai==1.46.1
pandas==2.0.3
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.3.0
seaborn==0.13.2
Werkzeug==2.2.3
